\name{group_cat}
\alias{group_cat}
\title{Create variables showing the means, sums et.c of categories}
\usage{
group_cat(x, 
          cat = c("category"),
          newvars = c("new_var_name"),
          vars = c("var_to_compute_from"),
          comp = sum
}
\description{
Bla bla.
}
\examples{
Bla bla}
